Moved presets decorators to core
