+++ core/src/autogluon/core/utils/decorators.py
@@ -1,22 +1,62 @@
-import time, logging
+import functools
+import logging
 
 logger = logging.getLogger(__name__)
 
-# decorator to calculate duration taken by any function. Logs times at debug level (logging level 10).
-def calculate_time(func):
-    # added arguments inside the inner1,
-    # if function takes any arguments,
-    # can be added like this.
-    def inner1(*args, **kwargs):
-        # storing time before function execution
-        begin = time.time()
 
-        output = func(*args, **kwargs)
+def unpack(g, *other_args):
+    """
+    Used to pass *args and **kwargs to a function g prior to entering function f via:
 
-        # storing time after function execution
-        end = time.time()
-        logger.debug("Total time taken in " + str(func.__name__)+": "+str(end - begin))
+    def g(a=0, **kwargs):
+        kwargs['b'] = a + 1
+        return **kwargs
 
-        return output
+    @unpack(g)
+    def f(**kwargs):
+        print(**kwargs)
 
-    return inner1
+    f(a=2)  # kwargs is now the output of g(a=2), which is {'b': 3}
+    f(c=4)  # kwargs is now the output of g(c=4), which is {'b': 1, 'c': 4}
+
+    """
+    def _unpack_inner(f):
+        @functools.wraps(f)
+        def _call(*args, **kwargs):
+            gargs, gkwargs = g(*other_args, *args, **kwargs)
+            return f(*gargs, **gkwargs)
+        return _call
+    return _unpack_inner
+
+
+def _apply_presets(preset_dict, *args, **kwargs):
+    """
+    Pair with `unpack` to alter input arguments with preset values.
+    """
+    if 'presets' in kwargs:
+        presets = kwargs['presets']
+        if presets is None:
+            return kwargs
+        if not isinstance(presets, list):
+            presets = [presets]
+        preset_kwargs = {}
+        for preset in presets:
+            if isinstance(preset, str):
+                preset_orig = preset
+                preset = preset_dict.get(preset, None)
+                if preset is None:
+                    raise ValueError(f'Preset \'{preset_orig}\' was not found. Valid presets: {list(preset_dict.keys())}')
+            if isinstance(preset, dict):
+                for key in preset:
+                    preset_kwargs[key] = preset[key]
+            else:
+                raise TypeError(f'Preset of type {type(preset)} was given, but only presets of type [dict, str] are valid.')
+        for key in preset_kwargs:
+            if key not in kwargs:
+                kwargs[key] = preset_kwargs[key]
+    return args, kwargs
+
+
+def apply_presets(preset_dict):
+    """Used as a decorator"""
+    return unpack(_apply_presets, preset_dict)
