+++ tabular/src/autogluon/tabular/configs/presets_configs.py
@@ -1,19 +1,6 @@
 
-import functools
-
-
-def unpack(g):
-    def _unpack_inner(f):
-        @functools.wraps(f)
-        def _call(*args, **kwargs):
-            gargs, gkwargs = g(*args, **kwargs)
-            return f(*gargs, **gkwargs)
-        return _call
-    return _unpack_inner
-
-
 # Dictionary of preset fit() parameter configurations.
-preset_dict = dict(
+tabular_presets_dict = dict(
     # Best predictive accuracy with little consideration to inference time or disk usage. Achieve even better results by specifying a large time_limit value.
     # Recommended for applications that benefit from the best possible model accuracy.
     best_quality={'auto_stack': True},
@@ -44,28 +31,3 @@ preset_dict = dict(
 
     # TODO: Consider HPO-enabled configs if training time doesn't matter but inference latency does.
 )
-
-
-def set_presets(*args, **kwargs):
-    if 'presets' in kwargs:
-        presets = kwargs['presets']
-        if presets is None:
-            return kwargs
-        if not isinstance(presets, list):
-            presets = [presets]
-        preset_kwargs = {}
-        for preset in presets:
-            if isinstance(preset, str):
-                preset_orig = preset
-                preset = preset_dict.get(preset, None)
-                if preset is None:
-                    raise ValueError(f'Preset \'{preset_orig}\' was not found. Valid presets: {list(preset_dict.keys())}')
-            if isinstance(preset, dict):
-                for key in preset:
-                    preset_kwargs[key] = preset[key]
-            else:
-                raise TypeError(f'Preset of type {type(preset)} was given, but only presets of type [dict, str] are valid.')
-        for key in preset_kwargs:
-            if key not in kwargs:
-                kwargs[key] = preset_kwargs[key]
-    return args, kwargs
